FIXED: testing fvalues before including in model
try_to_add_point was allowing the inclusion of points that lead to bad
function values. The interpolation was failing in such cases.

evaluate_new_fvalues now defaults to nan instead of 0. A bit safer in
case of failure.

Main algorithm is:
- testing every starting point for bad fvalues
- taking measures if trial points give bad fvalues

A few optimization problems are not failing anymore and are being
correctly solved as a result.